refactor: PathMapper 내부에서 static class 분리
- PathMapper 의 내부에 있던 Pair class 를 분리하고 Pair 를 사용해
  mapping 하고 있던 로직을 PathMapperFactory class 로 분리하였습니다.
