+++ src/main/java/webserver/WebServer.java
@@ -6,11 +6,12 @@ import java.net.Socket;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.handler.PathMapper;
+import webserver.handler.PathMapperFactory;
 
 public class WebServer {
     private static final Logger log = LoggerFactory.getLogger(WebServer.class);
 
-    private static final PathMapper pathMapper = new PathMapper();
+    private static final PathMapper pathMapper = PathMapperFactory.create();
     private static final int DEFAULT_PORT = 8080;
 
     public static void main(String args[]) throws Exception {
